move existing ModifierQuantization and tests to legacy file
